Revert "マイページのISR化"
This reverts commit 3aa488a0e75be8e29df24771b1fc5f344101bd79.
